requirements.txt
# AID: requirements.txt
# KT-v47 Core (Auditability, Logic, Math)
numpy>=1.26.4
pydantic>=2.7.1
scipy>=1.13.1
metric-temporal-logic>=0.4.0
streamlit>=1.30.0
pyyaml>=6.0.0
prometheus-client>=0.20.0
cryptography>=41.0.0
requests>=2.31.0

# Control Tower Dependencies (Human Sovereignty Interface)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Titanium X Protocol - God Level Upgrades (Pinned Versions)
z3-solver==4.12.2.0      # Formal verification (SMT solver)
merklelib==1.0           # Cryptographic ledger integrity
pysyncobj==0.3.14 ; python_version >= "3.11"   # Raft consensus (Py3.11 compatible)
pysyncobj==0.4.7  ; python_version <  "3.11"   # Raft consensus (older Pythons)
nemoguardrails==0.7.1    # Production-grade ethical filtering
pytest-asyncio>=0.23.0   # Async testing support
pytest-timeout>=2.2.0    # Timeout protection for tests
# fastembed>=0.2.0       # [DEPRECATED] Replaced by sentence-transformers for Python 3.14 stability
sentence-transformers>=2.3.0  # Production embedding backend (SentenceTransformer + ONNX)
onnxruntime>=1.16.0      # ONNX Runtime for optimized CPU inference
rapidfuzz>=6.2.0         # C++ accelerated fuzzy matching (Titanium Xv2)

# Apex Fine-Tuning Stack (ML Eng)
torch>=2.3.0
transformers>=4.40.0
accelerate>=0.30.0
peft>=0.11.0
bitsandbytes>=0.43.0
# -- Uncomment these if GPU and CUDA 12.1+ are available --
# unsloth[cu122]
# flash-attn>=2.6.0
